rename: form 폴더 inputs로 변경, button 폴더 buttons로 변경 #8
